+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this library will be documented in this file.
 
 Version in this change log adheres to [Semantic Versioning](http://semver.org/).
 
+### 3.0.4 (2020-12-01) [#48](https://github.com/Tinder/GitQuery/pull/48) Fix cli bug prefixing / to path ([tinder-aminghadersohi](https://github.com/tinder-aminghadersohi)
+
+### 3.0.3 (2020-12-01) [#48](https://github.com/Tinder/GitQuery/pull/48) Allow init-config to use includeGlobs in config ([tinder-aminghadersohi](https://github.com/tinder-aminghadersohi)
+    
+### 3.0.2 (2020-12-01) [#48](https://github.com/Tinder/GitQuery/pull/48) Bug Fix - Snakeyaml constructor call ([tinder-aminghadersohi](https://github.com/tinder-aminghadersohi)
+    - Upgrade to Snakeyaml 1.27 
+    - Simplify constructor call for Yaml to avoid this error: `org.yaml.snakeyaml.representer.Representer.<init>(Lorg/yaml/snakeyaml/DumperOptions;)V`
+    
+### 3.0.1 (2020-12-01) [#48](https://github.com/Tinder/GitQuery/pull/48) Bug Fix - init config ([tinder-aminghadersohi](https://github.com/tinder-aminghadersohi)
+    - Fix a bug related to the gradle plugin init config writing absolute path where relative is desired. 
+    
 ### 3.0.0 (2020-11-14) [#45](https://github.com/Tinder/GitQuery/pull/45) GitQuery 3 ([tinder-aminghadersohi](https://github.com/tinder-aminghadersohi)
     - New `init-config` command line parameter
     - New Gradle extension `gitQueryInit`